perf(agent): parallelize provider status detection
